Update API base URLs and refactor AllBlogsList component
