+++ client/src/api/useAxiosPublic.ts
@@ -1,7 +1,7 @@
 import axios, { AxiosInstance } from "axios";
 
 const axiosPublic: AxiosInstance = axios.create({
-  baseURL: "http://localhost:5000",
+  baseURL: "https://blog-website-server-theta.vercel.app",
 });
 
 const useAxiosPublic = () => {
