+++ client/src/api/useAxiosSecure.ts
@@ -5,7 +5,7 @@ import useAuth from "../hooks/useAuth";
 import toast from "react-hot-toast";
 
 const axiosSecure: AxiosInstance = axios.create({
-  baseURL: "http://localhost:5000",
+  baseURL: "https://blog-website-server-theta.vercel.app",
   withCredentials: true,
 });
